+++ include/wx/msw/wince/checklst.h
@@ -50,20 +50,20 @@ public:
                 const wxString& name = wxListBoxNameStr);
 
     // override base class virtuals
-    virtual void Delete(int n);
+    virtual void Delete(unsigned int n);
 
     // items may be checked
-    virtual bool IsChecked(size_t uiIndex) const;
-    virtual void Check(size_t uiIndex, bool bCheck = true);
+    virtual bool IsChecked(unsigned int uiIndex) const;
+    virtual void Check(unsigned int uiIndex, bool bCheck = true);
 
     // public interface derived from wxListBox and lower classes
     virtual void Clear();
-    virtual size_t GetCount() const;
+    virtual unsigned int GetCount() const;
     virtual int GetSelection() const;
     virtual int GetSelections(wxArrayInt& aSelections) const;
-    virtual wxString GetString(int n) const;
+    virtual wxString GetString(unsigned int n) const;
     virtual bool IsSelected(int n) const;
-    virtual void SetString(int n, const wxString& s);
+    virtual void SetString(unsigned int n, const wxString& s);
 
     // Implementation
     virtual bool MSWOnNotify(int idCtrl, WXLPARAM lParam, WXLPARAM *result);
@@ -73,12 +73,12 @@ protected:
 
     // protected interface derived from wxListBox and lower classes
     virtual int DoAppend(const wxString& item);
-    virtual void* DoGetItemClientData(int n) const;
-    virtual wxClientData* DoGetItemClientObject(int n) const;
-    virtual void DoInsertItems(const wxArrayString& items, int pos);
+    virtual void* DoGetItemClientData(unsigned int n) const;
+    virtual wxClientData* DoGetItemClientObject(unsigned int n) const;
+    virtual void DoInsertItems(const wxArrayString& items, unsigned int pos);
     virtual void DoSetFirstItem(int n);
-    virtual void DoSetItemClientData(int n, void* clientData);
-    virtual void DoSetItemClientObject(int n, wxClientData* clientData);
+    virtual void DoSetItemClientData(unsigned int n, void* clientData);
+    virtual void DoSetItemClientObject(unsigned int n, wxClientData* clientData);
     virtual void DoSetItems(const wxArrayString& items, void **clientData);
     virtual void DoSetSelection(int n, bool select);
     // convert our styles to Windows
